build admin bus list

+++ backend/src/services/admin.service.js
@@ -23,11 +23,19 @@ const createBus = async (req) => {
 };
 
 const deleteBusById = async (busId) => {
-  await prisma.buses.delete({
+  const bus = await prisma.buses.findUnique({
     where: {
       id: busId,
     },
   });
+
+  if (!bus) throw new ApiError(httpStatus.BAD_REQUEST, 'BUS NOT FOUND');
+  const result = await prisma.buses.delete({
+    where: {
+      id: busId,
+    },
+  });
+  if (!result) return false;
   return true;
 };
